update GetItem to also handle query with runID
