+++ pkg/debugapi/queue_partition.go
@@ -11,7 +11,6 @@ import (
 	pb "github.com/inngest/inngest/proto/gen/debug/v1"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"encoding/json"
 )
 
 var (
@@ -91,21 +90,4 @@ func (d *debugAPI) GetPartitionStatus(ctx context.Context, req *pb.PartitionRequ
 		Future:            int64(pt.Future),
 		Backlogs:          int64(pt.Backlogs),
 	}, nil
-}
-
-func (d *debugAPI) GetQueueItem(ctx context.Context, req *pb.QueueItemRequest) (*pb.QueueItemResponse, error) {
-	queueItem, err := d.queue.ItemByID(ctx, req.GetId())
-	if err != nil {
-		if errors.Is(err, redis_state.ErrQueueItemNotFound) {
-			return nil, status.Error(codes.NotFound, "not found")
-		}
-		return nil, status.Error(codes.Unknown, fmt.Errorf("error retrieving queue item: %w", err).Error())
-	}
-
-	byt, err := json.Marshal(queueItem)
-	if err != nil {
-		return nil, status.Error(codes.Internal, fmt.Errorf("error marshalling queue item: %w", err).Error())
-	}
-
-	return &pb.QueueItemResponse{Data: byt}, nil
 }
